+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.1] - 2024-05-20
 
 ### Added
-- Added a new class (PdfGenerator) for generating PDF documents from HTML content using Android's WebView and PrintManager
-- Introduced a new class (HtmlPopulator) to populate HTML templates with data from a Questionnaire Response
+- Added the in-app PDF Generation feature
+  1. Added a new class (PdfGenerator) for generating PDF documents from HTML content using Android's WebView and PrintManager
+  2. Introduced a new class (HtmlPopulator) to populate HTML templates with data from a Questionnaire Response
+  3. Implemented functionality to launch PDF generation using a configuration setup
 
 ## [1.1.0] - 2024-02-15